+++ CHANGELOG.md
@@ -84,6 +84,7 @@
 - Plugin menu-links: Add option to show the select link and other repeated links only on hover
 - New plugin: Display images in select, same as in Adminer Editor
 - New plugin: Log in by a passkey holding the passwords
+- New plugin: Warn by a red strip if Adminer or the database doesn't run on the local machine
 ### Internal
 - Register JavaScript event handlers by a data attribute instead of a script element
 - Register the event handlers in functions.js instead of an inline script